Move CallSiteProfile for cross-module access and cleanup
